[BE] refactor: HearitWithPlayTime 인터페이스명 변경 #487
